Physics
Physics testing.
There are different Look classes for reasons of physics movements, e.g.
the camera should only rotate up/down and the player left/right.
The player will bounce a bit, barely be able to look up and down, not be
able to move at all, and eventually fall through the plane.

I am trying to fix this...still no idea. What I need help on is the
glitch of falling through the terrain and also on how to apply positions
and rotations to JBullet from Java code without applying rigidbody
forces.