GameObject improvements, formatting, etc
I added names to GameObjects and the ability to get the first Component
it comes across of a certain type (allowing to find a component in a GO
like a collider) along with finding children by name. I also added the
ability to set Collider mass in the game. Lastly, I formatted some
stuff.