Fixed Scene Loading
Now works properly. The Camera was added twice, once as a child to the
player and another as if it were a GameObject in the root of the scene.
To fix this I made sure that it checked if a Node's parent node was the
main list before adding it to the root.